moved all functions to utils.js
