+++ front_end/src/components/PrivateRoute.js
@@ -1,6 +1,6 @@
 import React, { useEffect, useState } from 'react';
 import { Route, Redirect } from 'react-router-dom';
-import { isLogin } from '../utils/login';
+import { isLogin } from '../utils/utils';
 
 const PrivateRoute = ({component: Component, restricted, ...rest}) => {
     const [isMounted, setMounted] = useState(false);
