+++ front_end/src/views/HomePage.js
@@ -1,8 +1,6 @@
 import React from "react";
 import "./HomePage.css";
-import Cookies from 'universal-cookie';
-import getBooks from '../utils/homepage';
-import { getGenres } from '../utils/labels'; 
+import { getGenres, getBooks } from '../utils/utils'; 
 import NavigationBar from '../components/NavigationBar'
 import Container from '@material-ui/core/Container'
