+++ front_end/src/views/Login.js
@@ -3,7 +3,7 @@ import TextField from "@material-ui/core/TextField";
 import Button from "@material-ui/core/Button";
 import { useHistory } from "react-router-dom";
 import { makeStyles } from "@material-ui/core/styles";
-import { login }from "../utils/login";
+import { login }from "../utils/utils";
 import Cookies from 'universal-cookie';
 import "./Login.css";
