+++ front_end/src/views/Themes.js
@@ -4,7 +4,7 @@ import { useHistory } from "react-router-dom";
 import { makeStyles } from "@material-ui/core/styles";
 import "./Themes.css";
 import labels from "../static/labels"
-import { saveThemes } from "../utils/labels"
+import { saveThemes } from "../utils/utils"
 
 
 const ThemesView = props => {
